refactor: update event schema to match Go v2 and rename fields for consistency

+++ zitadel-backend/src/lib/database/migrations/002_01_create_projection_states_table.sql
@@ -4,7 +4,7 @@
 
 CREATE TABLE IF NOT EXISTS projection_states (
     name VARCHAR(255) PRIMARY KEY,
-    position BIGINT NOT NULL DEFAULT 0,
+    position DECIMAL NOT NULL DEFAULT 0,
     last_processed_at TIMESTAMP WITH TIME ZONE,
     status VARCHAR(50) NOT NULL DEFAULT 'stopped',
     error_count INTEGER NOT NULL DEFAULT 0,
